Integration tests: Allow to run a specific test from tox
Previously the path of the integration tests was specified explicitly
in tox.ini and we cannot run a specific test from tox.
{posargs} existed in the previous command line, but it can be used
only to specify extra options and we cannot change the test path,
so the whole integration tests are always run.
It can be solved by passing the integration test path as
the default value of {posargs}.

The most possible case where we would like to change the test command
is to run a specific set of tests instead of the whole tests,
so I believe this solution would work.

Change-Id: Ifa478a360a8af8b3d959de7ec632eedef484391d