Drop py36 and py37 support
Python runtimes for Zed are 3.8, 3.9 and optionally 3.10 [1][2],
so tox envs in tox.ini are updated accordingly.

Declare python_requires >=3.8 in setup.cfg
python_requires are declared in many projects during dropping
python 2.7 support. It is good to follow the convention and
it also clarifies our python interpreter requirement.

This commit also cleans up python2.7 related stuffs in setup.py.
Both are related to the supported versions of python,
so I made both changes in a single commit.

[1] https://governance.openstack.org/tc/reference/runtimes/zed.html
[2] https://opendev.org/openstack/openstack-zuul-jobs/src/commit/b740b42fb199d4d695391aaaddf46ca55eab79ac/zuul.d/project-templates.yaml#L542

Change-Id: I4161816fd1fec19a3834cc7d994b3f761e1ab2ce
Closes-Bug: #1960887