Speed up integration tests
We have a place in the form-handling code where we look for the
blue form submit button, and if it's not found, we look for the
red form submit button. It looks correct at first glance, however
it takes 10 seconds to throw NoSuchElementException, because there
is a delay that makes sure that the page loads fully before raising
that exception. Replacing this with a selector that finds both the
blue and the red buttons in one query speeds the whole integration
tests suite by half an hour, because there is no more waiting on
every delete confirmation dialog.

Change-Id: Ie256118c0a84a8868393018b25b0aa049582a17d
(cherry picked from commit 2e6eca294718deb7e82f50c071d332377c35085f)