Remove console_type parameter for server_mks_console function
The "get_mks_console" use "console_type" as param, but causes an issue on Nova.
In horizon logs we have "Recoverable error: No available console found." and in the dashboard we have "Unable to load console. Please reload page to try again." when we load the console.
There is no need to call the function with this parameter, as it's already defined automatically since microversion 2.53.

Change-Id: I776b19f053ca74c699ca069e04553740f1e83b3e