List ports with project_id.
During launch_instance, all ports belonging to network are shown if the
network is shared. However each port actually has a 'project_id'
property, and the port cannot be used for any instance outside of that
project. When a user attempts to start an instance with a port from a
different project, it fails like this: nova.exception.PortNotUsable

Fix by filters port_list with project_id.

Closes-bug: #1996108
Change-Id: If94e91e5ae6568e06096380c2bf65d08d20a5958