Set Nodeset to debian-bullseye for npm, integration and selenium job
This patch changes nodeset to "debian-bullseye" from "ubuntu-focal" for
horizon-selenium-headless, horizon-nodejs, and horizon-integration-tests
jobs. Debian bullseye is also a runtime for the 2023.1 cycle as
mentioned here [1]. Right now if we change nodeset to "ubuntu-jammy"
all of the above jobs start failing. One possible solution to run these
jobs on "ubuntu-jammy" nodeset is to install firefox as a deb. package
instead of using snap as discussed here [2]. We can switch nodeset to
"ubuntu-jammy" in the future once the firefox issue with snap is fixed [3].

Also geckodriver version is upadted to latest(i.e. v0.32.0) to include
latest bug fix.

[1] https://governance.openstack.org/tc/reference/runtimes/2023.1.html
[2] https://meetings.opendev.org/irclogs/%23openstack-infra/%23openstack-infra.2022-11-23.log.html#t2022-11-23T14:15:08
[3] https://bugs.launchpad.net/ubuntu/+source/snapd/+bug/1951491

Depends-On: https://review.opendev.org/c/zuul/zuul-jobs/+/865459

Related-Bug: #1996638

Change-Id: Ic50fbb8b29fc5dec42d1a5e94095c9777c7d5fb6