Remove Description from Edit Image Metadata
Image description can be edited through Edit Image -> Image Detail,
however it also appears as a property in Edit Image -> Metadata when
the description is not empty. It's confusing to have it in two sections
of one dialog form, and moreover, the Image Detail field always
overwrites the Metadata field.

This patch removes Description from Edit Metadata.

Closes-Bug: #2017698

Change-Id: I041948b4b71f00894f03cc30cfa4592df4a100c3