Make horizon-integration-tests job voting again
This patch makes "horizon-integration-tests" job voting
again. It was changed to non-voting so that patch that
update Xstatic-jQuery upper bound, update cirros version
for the horizon-integration job and a few other changes
which are required to fix the horizon-integration can get merged.

The patch [1] is already merged and the upper bound in
openStack/requirement also got updated[2]. So let's make
horizon-integration job voting again.

[1] https://review.opendev.org/c/openstack/horizon/+/887548
[2] https://review.opendev.org/c/openstack/requirements/+/887933

Change-Id: I0059c2cd47e6b540306337caa3e0f826f3b755a9