i18n: Avoid using variables in gettext() in network QoS edit rule
If an argument to gettext() contains variables, strings extracted
and strings used when rendered will be different (as variables are
evaluated as empty during string extraction to POT).
As a result, translations for these strings aree never used.

This commit fixes the issue by defining strings with variable
substitions (like%(id)s) in gettext() combined with interpolate().

Labels of "Rule to edit" pulldown menu in "Edit Rule" form are
updated to match the format used in "Delete Rule" form.

Closes-Bug: #2037278
Change-Id: I06cb56ada2c842e5a445026b4649260bf9893841