pytest-based selenium tests add fix recording issue
When tests are very short and passed (they do not wait for any element).
Specifically: test_user_login, test_admin_login.
The action in UI (login and check the homepage) is super quick (0.5s-1s).
So the video is not even created - we do not need to clean it up.

Change-Id: Id8433842247258f472fa3951fcdfa2b0e78ae781