Drop pymongo from requirements
The pymongo library is required only when profiler middleware is
enabled. The middleware is meant for development (The module path
contains developer) and is not for production.

Also it's not directly imported but used via osprofiler. In case
a different backend is used in osprofiler then the library is not at
all used. Users should install the appropriate library according to
the backend actually used, based on the extra dependencies of
osprofiler.

Change-Id: I1258a3e3c86c98a3b35ee413ae113f1eb7dcf00b