Make horizon-tox-bandit-baseline job non-voting
This patch makes "horizon-tox-bandit-baseline" job
non-voting becuase due to some changes in latest
version of bandit it starts failing. This issue is already
fixed in horizon master barnch and backported till stable/2023.1
branch but if we apply the same fix in stable/zed branch
then package dependency issue happens.
So it's better to make this job non-voting.

Change-Id: I3edfd885ff582791a7bba6cd03d301852fac8a17