Add help text for edit instance form
This commit adds help text to the Edit Instance form to describe the
limitations of the text allowed in the name and description. The help text guides the maximum length for the instance name and description
and advises against using special characters or leading or trailing spaces. By providing this information, users will be better informed when modifying instance details, reducing the likelihood of
encountering errors.

Closes-Bug: #1981165
Change-Id: If8879c20b2842c3dd769e4cdef80834219c637cd