Force scope when listing domains
since Caracal, when using domain-scoped token, keystone only returns
the domain the token is scoped to when listing domains [0].

Since Horizon does some behind-the-scenes swap of token scope when
doing some requests to Keystone, this breaks the Identity->Domains panel
for admins.

This patch forces the domain_list call to always use the original
auth scope, w/o a swap to the domain-scoped token.

[0] https://review.opendev.org/c/openstack/keystone/+/900028

Closes-Bug: #2067075
Change-Id: I4ff5f2de01c0bb13cfbb5136f40afe8187135686