Delete container show duplicate toast notifications
When trying to delete a container with objects in it the
Swift rest API rejects the request with 409.

In the JS code we catch() the promise from the request
and show a toast error that the request failed but we
don't throw in that catch() so the then() callback that
the callee specified gets called which shows a toast
message that the container delete was successful and
removes it from the list, if you refresh the page the
container is still in the list (as expected).

This subtle error probably exists in multiple places
where the same kind of logic exists.

Change-Id: Iadf1e58f23c935a7dd599b9fd3444a0a22396213