Update job nodesets from bullseye to bookworm
This change updates the nodesets to use debian bookworm
in order to provide a newer libvirt version.

https://review.opendev.org/c/openstack/devstack/+/922630 enables
a number of optimisations to the base devstack job to improve
performance and reduce memory pressure in order to avoid timeouts
and improve the stability of the CI jobs.

One of the optimizations involved limiting the qemu block translation
cache size to 128mb which required libvirt 8.0.0. Bullseye does
not provide that but bookworm does.

This change updates Horizon to use bookworm to accommodate this overall
change in job configuration. Note Nova will raise its min libvirt version
to 8.0.0 in 2025.1 so this change would be required in anycase next
cycle.

Change-Id: I42eb5778c861515b4984f762bdf03d2c90d5da41