Sort image source choices by name for volume
When creating a volume and selecting to use image as a
source we get a dropdown of the images that we can select
but this list is not sorted based on image name.

This changes so that it's sorted by the image name
ascending to match the instance launch dialog that
already lists images in ascending order by name.

Change-Id: I17212e460e307e08b2b6c2a8a68a14ffde8cc04b