Update horizon tests to use the RBAC new roles
Horizon openstack dashboard rest API unit tests only
pass the project_id to oslo.policy to vrify the service
policy rule RBAC. This was passing till now as services
allow 'owner' (which only check project_id) to pass
the policy checks.

As per new RBAC, project_id is not enough and proper
role should be passed. For example 'owner' means in new
RBAC is either 'member' or 'reader' role with project_id.

oslo.policy 4.4.0 enable the new RBAC by default
- https://review.opendev.org/c/openstack/releases/+/925032

and requirement change to use the oslo.policy 4.4.0 in
upper-constraints is blocked with the horizon failure
- https://review.opendev.org/c/openstack/requirements/+/925464

This commit fixes the test to use the right role along with
the project_id so that policy rules can be checked correctly.

Needed-By: https://review.opendev.org/c/openstack/requirements/+/925464
Change-Id: I840996fd2635bca853da02c630a7ab3761576821