Drop USE_L10N
This option was deprecated in django 4.0 and was removed in 5.0 .

https://docs.djangoproject.com/en/5.1/releases/4.0/#localization
```
In order to follow good practice, the default value of the USE_L10N
setting is changed from False to True.

Moreover USE_L10N is deprecated as of this release. Starting with
Django 5.0, by default, any date or number displayed by Django will be
localized.
```

Change-Id: I81e1c790fe8df7851f1f91c0851d64af17c7dbdb