Do column name translations at runtime, not at import
The column's verbose_name was forced to a string from a lazy translation
at the object's creation time, which happens when the table definitions
are imported, and when the user language preference is not set yet.

Instead force it to a string at the very last moment, when it is used in
a template.

Fixes-bug: 2093276
Change-Id: I9cd3bc2535915a858aa380287fccd89d0f19cf7d