Catch callback in network service doesn't throw
This is the same issue as described in [1].

Example: When octavia-dashboard uses the network service
to associate a floating IP and the call fails the then()
callback saying it was successful will be called in
octavia-dashboard alongside the error toast that is
shown in the catch() callback.

We should throw in the error callback so that the
then() callback is not called.

[1] https://review.opendev.org/c/openstack/horizon/+/922094

Change-Id: Ieb4b0b76b59fca2e750c576725b0a8987b0837a1
Signed-off-by: Tobias Urdin <tobias.urdin@binero.com>