Image Sharing - Visibility - fix for long phrases
Change include both forms - create image and edit image

At this moment, in block Image Sharing, buttons in Visibility function displayed without problem only for English language, if start using languages ​​with long phrases - buttons are transferred to another line, which does not look very nice.

Overall, I think we can change the design of this element to a drop-down list, but given that this is an angular form, I think it would be better to do this when back to python.

Change-Id: Iefd92c8abf52372d74e64f601fb9557006f0a876