Add changing instance password support
This brings the CLI password changing command (nova set-password
<instance_id>) into Horizon.

The action is allowed if the instance is in one of the active states or
is shuttof and is enabled in settings.

A modal is presented for entering the new password and after
confirmation, the nova client is called.

Based on MR: I550403490094ab040801bd7de9f7cd1f20518adb

Change-Id: I039d273cb10d532dfb2d18ef8e8943946ded9da6
Co-authored-by: Sergiu Miclea <smiclea@cloudbasesolutions.com>
Signed-off-by: Dmitriy Chubinidze <dcu995@gmail.com>