pytest-based selenium test wait for angular timeout
Timeout for script for waiting on angular to be ready seems
to be not enough sometimes (10s).
As this async script is working as checking the state of
angular, increasing the timeout should not bring any
negative effects, only positive.
Now the timeout for the script will be dependent on
the page timeout variable.

Change-Id: I32067bf1451b48536a5ca96e5cee24a4b80158c2
Signed-off-by: Jan Jasek <jjasek@redhat.com>