Add integration tests for region selection and switching
Adds integration tests that ensure that the correct region selection
is present on the login screen, and that logging in with each of the
regions selected results in switching to that region.

Also adds a test that uses the region switching menu to switching
between all available regions.

The tests are skipped if AVAILABLE_REGIONS is empty.

Change-Id: I46ec58e17496933677302789ee8551c724a37b10
Signed-off-by: Radomir Dopieralski <openstack@dopieralski.pl>