Don't send enable_port_security when disallowed by policy
When a user creates a network port, if they don't have the rights
to change port security, they will be unable to submit the form.
The solution is to not send any value for port security when the
user doesn't have the rights to change it.

Change-Id: I70d15b71083c3934ed48f24765b42a62daf58cf8
Signed-off-by: Dong Ma <dong.ma@vexxhost.com>
(cherry picked from commit 4b933df52ae64314b26c2d8028203b70091736c9)