Link to SCSS styles instead of inlining them in every page
For quite some time now we have been inlining all the SCSS compiled
styles on every page of Horizon, as a workaround to preserve namespace.
This is less than optimal, as it generates kilobytes of additional
data on every page load, and it prevent us from reusing the styles
with shadow dom and web components.

This patch is to fix this, so that we link to
the stylesheets instead of inlining them. It will require some changes
in how the stylesheets are handled.

Signed-off-by: Radomir Dopieralski <openstack@dopieralski.pl>
Change-Id: Ia64b9f8a06a6c413aec2d62252a39843b609c305